requirements.txt
transformers
openai
fastapi
uvicorn
pydantic
python-dotenv
streamlit
torch
